feat(): refactor validation
